requirements.txt
ipython==8.12.0
matplotlib==3.7.1
numpy==1.21.5
opencv_python==4.7.0.72
Pillow==9.5.0
python==3.9.12
scikit_learn==1.2.2
seaborn==0.12.2
torch==2.0.0
torchmetrics==0.11.0
torchsummary==1.5.1
torchvision==0.13.1
